// src/types.ts
import { User } from '@privy-io/react-auth';

export type BountySource = 'twitter' | 'lens' | 'bounty bird' | 'farcaster';

export enum BountyStatus {
	OPEN = 'OPEN',
	CLOSED = 'CLOSED'
}

export interface IClaim {
	user_id: string;
	created_at: Date;
}

export enum EReactionEmoji {
	THUMBS_UP = '👍',
	THUMBS_DOWN = '👎'
}

export interface IReaction {
	user_id: string;
	created_at: Date;
	reaction: EReactionEmoji;
}

export interface IBountyReply {
	id: string; //source_id is the id here
	created_at: Date;
	updated_at: Date;
	source_author_id: string;
	display_name: string;
	username: string;
	text: string;
	source: BountySource;
	txHash?: string;
	accepted?: boolean;
	deleted?: boolean;
}

export interface IBounty {
	id: string;
	username: string;
	display_name: string;
	status: BountyStatus;
	source: BountySource;
	source_author_id?: string;
	source_id?: string;
	source_text?: string;
	task: string;
	amount: string;
	deadline?: Date;
	max_claims: number;
	created_at: Date;
	updated_at: Date;
	claims: IClaim[];
	replies?: IBountyReply[];
	reactions?: IReaction[];
	replies_count?: number;
	deleted?: boolean;
}

export interface IApiResponse<T> {
	data: T | null;
	error: string | null;
	status: number;
}

export interface IApiErrorResponse {
	error: string;
}

export interface IApiSuccessResponse {
	message: string;
}

export interface IBountyListingResponse {
	totalCount: number;
	bounties: IBounty[];
}

export interface IUserAddress {
	address: string;
	wallet?: string;
	created_at: Date;
	updated_at: Date;
}

export interface IUser {
	bio: string | null;
	created_at?: Date;
	id: string;
	img_url: string | null;
	is_username_autogenerated: boolean | null;
	lens_handle?: string | null;
	name: string | null;
	privy_user?: User | null;
	username: string | null;
	updated_at?: Date;
	wishlist?: string[];
}

export interface ISocialVerificationStartResponse {
	code: string;
}

export interface ISocialVerificationEntry {
	code: string;
	handle: string;
}

export enum NotificationStatus {
	SUCCESS = 'success',
	ERROR = 'error',
	WARNING = 'warning',
	INFO = 'info'
}